+++ packages/GitWrapper/src/GitWorkingCopy.php
@@ -396,6 +396,8 @@ final class GitWorkingCopy
 
     /**
      * @code $git->apply('the/file/to/read/the/patch/from');
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function apply(...$argsAndOptions): string
     {
@@ -405,6 +407,8 @@ final class GitWorkingCopy
     /**
      * @code $git->bisect('good', '2.6.13-rc2');
      * $git->bisect('view', ['stat' => true]);
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function bisect(string $subCommand, ...$argsAndOptions): string
     {
@@ -415,6 +419,8 @@ final class GitWorkingCopy
     /**
      * @code $git->branch('my2.6.14', 'v2.6.14');
      * $git->branch('origin/html', 'origin/man', ['d' => true, 'r' => 'origin/todo']);
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function branch(...$argsAndOptions): string
     {
@@ -423,6 +429,8 @@ final class GitWorkingCopy
 
     /**
      * @code $git->checkout('new-branch', ['b' => true]);
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function checkout(...$argsAndOptions): string
     {
@@ -448,6 +456,8 @@ final class GitWorkingCopy
      *
      * @code $git->commit('My commit message');
      * $git->commit('Makefile', ['m' => 'My commit message']);
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function commit(...$argsAndOptions): string
     {
@@ -466,6 +476,8 @@ final class GitWorkingCopy
      *
      * @code $git->config('user.email', 'opensource@chrispliakas.com');
      * $git->config('user.name', 'Chris Pliakas');
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function config(...$argsAndOptions): string
     {
@@ -475,6 +487,8 @@ final class GitWorkingCopy
     /**
      * @code $git->diff();
      * $git->diff('topic', 'master');
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function diff(...$argsAndOptions): string
     {
@@ -486,6 +500,8 @@ final class GitWorkingCopy
      *
      * @code $git->fetch('origin');
      * $git->fetch(['all' => true]);
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function fetch(...$argsAndOptions): string
     {
@@ -496,6 +512,8 @@ final class GitWorkingCopy
      * Print lines matching a pattern.
      *
      * @code $git->grep('time_t', '--', '*.[ch]');
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function grep(...$argsAndOptions): string
     {
@@ -520,6 +538,8 @@ final class GitWorkingCopy
      *
      * @code $git->log(['no-merges' => true]);
      * $git->log('v2.6.12..', 'include/scsi', 'drivers/scsi');
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function log(...$argsAndOptions): string
     {
@@ -528,7 +548,7 @@ final class GitWorkingCopy
 
     /**
      * Join two or more development histories together.
-     *
+     * @param mixed ...$argsAndOptions
      * @code $git->merge('fixes', 'enhancements');
      */
     public function merge(...$argsAndOptions): string
@@ -540,8 +560,7 @@ final class GitWorkingCopy
      * Move or rename a file, a directory, or a symlink.
      *
      * @code $git->mv('orig.txt', 'dest.txt');
-     * @param string $source The file / directory being moved.
-     * @param string $destination The target file / directory that the source is being move to.
+     *
      * @param mixed[] $options
      */
     public function mv(string $source, string $destination, array $options = []): string
@@ -554,6 +573,8 @@ final class GitWorkingCopy
      * Fetch from and merge with another repository or a local branch.
      *
      * @code $git->pull('upstream', 'master');
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function pull(...$argsAndOptions): string
     {
@@ -562,8 +583,9 @@ final class GitWorkingCopy
 
     /**
      * Update remote refs along with associated objects.
-     *
      * @code $git->push('upstream', 'master');
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function push(...$argsAndOptions): string
     {
@@ -572,8 +594,9 @@ final class GitWorkingCopy
 
     /**
      * Forward-port local commits to the updated upstream head.
-     *
      * @code $git->rebase('subsystem@{1}', ['onto' => 'subsystem']);
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function rebase(...$argsAndOptions): string
     {
@@ -584,6 +607,8 @@ final class GitWorkingCopy
      * Manage the set of repositories ("remotes") whose branches you track.
      *
      * @code $git->remote('add', 'upstream', 'git://github.com/cpliakas/git-wrapper.git');
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function remote(...$argsAndOptions): string
     {
@@ -594,6 +619,8 @@ final class GitWorkingCopy
      * Reset current HEAD to the specified state.
      *
      * @code $git->reset(['hard' => true]);
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function reset(...$argsAndOptions): string
     {
@@ -605,10 +632,6 @@ final class GitWorkingCopy
      *
      * @code $git->rm('oldfile.txt');
      *
-     * @param string $filepattern Files to remove from version control. Fileglobs (e.g.  *.c) can be given to add all
-     * matching files. Also a leading directory name (e.g. dir to add dir/file1 and dir/file2) can be given to add all
-     * files in the directory, recursively.
-     *
      * @param mixed[] $options
      */
     public function rm(string $filepattern, array $options = []): string
@@ -618,8 +641,6 @@ final class GitWorkingCopy
     }
 
     /**
-     * Show various types of objects.
-     *
      * @code $git->show('v1.0.0');
      *
      * @param string $object The names of objects to show. For a more complete list of ways to spell
@@ -637,6 +658,8 @@ final class GitWorkingCopy
      * Show the working tree status.
      *
      * @code $git->status(['s' => true]);
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function status(...$argsAndOptions): string
     {
@@ -647,6 +670,8 @@ final class GitWorkingCopy
      * Create, list, delete or verify a tag object signed with GPG.
      *
      * @code $git->tag('v1.0.0');
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function tag(...$argsAndOptions): string
     {
@@ -657,6 +682,8 @@ final class GitWorkingCopy
      * Remove untracked files from the working tree
      *
      * @code $git->clean('-d', '-f');
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function clean(...$argsAndOptions): string
     {
@@ -667,6 +694,8 @@ final class GitWorkingCopy
      * Create an archive of files from a named tree
      *
      * @code $git->archive('HEAD', ['o' => '/path/to/archive']);
+     *
+     * @param mixed ...$argsAndOptions
      */
     public function archive(...$argsAndOptions): string
     {