+++ packages/PackageBuilder/src/Parameter/ParameterProvider.php
@@ -32,6 +32,9 @@ final class ParameterProvider
         return $this->parameters[$name] ?? null;
     }
 
+    /**
+     * @param mixed $value
+     */
     public function changeParameter(string $name, $value): void
     {
         $this->parameters[$name] = $value;