+++ packages/TokenRunner/src/ReflectionDocBlock/DocBlockSerializerFactory.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Symplify\TokenRunner\DocBlock;
+namespace Symplify\TokenRunner\ReflectionDocBlock;
 
 use PhpCsFixer\WhitespacesFixerConfig;
 use phpDocumentor\Reflection\DocBlock\Serializer;
@@ -11,7 +11,6 @@ use phpDocumentor\Reflection\DocBlock\Serializer;
  * - that adds spaces after empty tag
  * - that adds "\" prefix to every type
  */
-
 final class DocBlockSerializerFactory
 {
     public static function createFromWhitespaceFixerConfigAndContent(