+++ packages/TokenRunner/src/Wrapper/FixerWrapper/DocBlockWrapper.php
@@ -12,12 +12,12 @@ use phpDocumentor\Reflection\DocBlock\Serializer;
 use phpDocumentor\Reflection\DocBlock\Tags\Param;
 use phpDocumentor\Reflection\DocBlock\Tags\Return_;
 use phpDocumentor\Reflection\DocBlock\Tags\Var_;
-use phpDocumentor\Reflection\DocBlockFactory;
 use phpDocumentor\Reflection\Types\Array_;
 use phpDocumentor\Reflection\Types\Compound;
 use Symplify\TokenRunner\DocBlock\ArrayResolver;
-use Symplify\TokenRunner\DocBlock\DocBlockSerializerFactory;
 use Symplify\TokenRunner\Guard\TokenTypeGuard;
+use Symplify\TokenRunner\ReflectionDocBlock\CleanDocBlockFactory;
+use Symplify\TokenRunner\ReflectionDocBlock\DocBlockSerializerFactory;
 
 final class DocBlockWrapper
 {
@@ -66,11 +66,9 @@ final class DocBlockWrapper
             $this->docBlock = new DocBlock($token->getContent());
         }
 
-        $docBlockFactory = DocBlockFactory::createInstance();
-
         $content = $token ? $token->getContent() : $docBlock->getContent();
 
-        $this->phpDocumentorDocBlock = $docBlockFactory->create($content);
+        $this->phpDocumentorDocBlock = (new CleanDocBlockFactory())->createFromContent($content);
         $this->originalContent = $content;
     }